Move MTRPS tests into RPS tests
