+++ tests/engine_tests/rockpaperscissors_test.py
@@ -4,16 +4,60 @@ from battle_engine.rockpaperscissors import RPSEngine
 from agent.rps_agent import RPSAgent
 
 
-def test_run_game():
-    """Test RPS Game executes properly."""
-    rps_engine = RPSEngine(bias=1)
-    rps_paper = RPSAgent(strategy_in="paper")
-    rps_rock1 = RPSAgent(strategy_in="rock")
-    rps_rock2 = RPSAgent(strategy_in="rock")
+def test_param_validation():
+    """Test that invalid parameters are caught."""
+    try:
+        RPSEngine(num_games=2)
+        assert False
+    except AttributeError:
+        pass
 
-    assert rps_engine.run(rps_paper, rps_rock1) == 1
-    assert rps_engine.run(rps_rock1, rps_paper) == 0
-    assert rps_engine.run(rps_rock1, rps_rock2) == 1
+    try:
+        RPSEngine(num_games=-1)
+        assert False
+    except AttributeError:
+        pass
+
+    try:
+        RPSEngine(num_games=0)
+        assert False
+    except AttributeError:
+        pass
 
 
-test_run_game()
+def test_basic_results(RPS):
+    """
+    Test that RPS results are still valid.
+
+    :param RPS: RPSEngine
+        Engine to use for this test.
+    """
+    rock_player = RPSAgent(strategy_in="rock")
+    paper_player = RPSAgent(strategy_in="paper")
+    scissors_player = RPSAgent(strategy_in="scissors")
+
+    # Rock beats Scissors
+    rs_outcome = RPS.run(rock_player, scissors_player)
+    assert rs_outcome == 1
+    # Rock loses to Paper
+    rp_outcome = RPS.run(rock_player, paper_player)
+    assert rp_outcome == 0
+
+    # Paper beats Rock
+    pr_outcome = RPS.run(paper_player, rock_player)
+    assert pr_outcome == 1
+    # Paper loses to Scissors
+    ps_outcome = RPS.run(paper_player, scissors_player)
+    assert ps_outcome == 0
+
+    # Scissors beats Paper
+    sp_outcome = RPS.run(scissors_player, paper_player)
+    assert sp_outcome == 1
+    # Scissors loses to Rock
+    sr_outcome = RPS.run(scissors_player, rock_player)
+    assert sr_outcome == 0
+
+
+test_basic_results(RPSEngine())
+test_basic_results(RPSEngine(num_games=3))
+test_param_validation()
